Feat: Added verifyKeyPage to verify the given keys
