+++ src/Screens/HashVerificationPage.css
@@ -1,6 +1,6 @@
 .hash-field {
     width: 100%;
-    margin-bottom: 20px;
-    padding: 10px;
-    margin-right: 10px;
+    margin-bottom: 1.5rem;
+    padding: 0.75rem;
+    text-align: center;
 }
