+++ src/Screens/VerifyKeyPage.css
@@ -1,0 +1,10 @@
+.hash-field {
+  width: 100%;
+  margin-bottom: 1.5rem;
+  padding: 0.75rem;
+  text-align: center;
+}
+
+.verify-button {
+  margin-top: 5rem;
+}
